# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão de empresas, pesquisas, dashboards, auditoria e submissões anônimas.\n\nFluxo sugerido para clientes frontend:\n1) Autenticar em /auth/login\n2) Enviar token JWT no header Authorization: Bearer \u003ctoken\u003e\n3) Consumir endpoints protegidos em /api/v1\n",
        "title": "API de Pesquisa de Clima Organizacional",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Equipe Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/empresas/{empresa_id}/setores": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setores"
                ],
                "summary": "Listar setores por empresa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da empresa",
                        "name": "empresa_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/empresas/{empresa_id}/setores/nome/{nome}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setores"
                ],
                "summary": "Buscar setor por nome",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da empresa",
                        "name": "empresa_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nome do setor",
                        "name": "nome",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/setores": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setores"
                ],
                "summary": "Criar setor",
                "parameters": [
                    {
                        "description": "Dados do setor",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto.SetorCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/setores/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setores"
                ],
                "summary": "Buscar setor por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do setor",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setores"
                ],
                "summary": "Atualizar setor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do setor",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto.SetorUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setores"
                ],
                "summary": "Remover setor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do setor",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "organizational-climate-survey_backend_internal_application_dto.SetorCreateRequest": {
            "type": "object",
            "required": [
                "id_empresa",
                "nome_setor"
            ],
            "properties": {
                "descricao": {
                    "description": "Descrição opcional, com limite de tamanho",
                    "type": "string",
                    "maxLength": 500,
                    "example": "Setor responsável por gestão de pessoas"
                },
                "id_empresa": {
                    "description": "Identificador da empresa associada",
                    "type": "integer",
                    "example": 1
                },
                "nome_setor": {
                    "description": "Nome do setor (obrigatório e limitado)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Recursos Humanos"
                }
            }
        },
        "organizational-climate-survey_backend_internal_application_dto.SetorUpdateRequest": {
            "type": "object",
            "properties": {
                "descricao": {
                    "description": "Nova descrição (opcional)",
                    "type": "string",
                    "maxLength": 500,
                    "example": "Área estratégica de gestão de talentos"
                },
                "nome_setor": {
                    "description": "Novo nome do setor (opcional)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "People Ops"
                }
            }
        },
...
        "organizational-climate-survey_backend_internal_application_dto_response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Codigo de erro para consumo de clientes",
                    "type": "string",
                    "example": "BAD_REQUEST"
                },
                "data": {
                    "description": "Dados retornados, genéricos"
                },
                "error": {
                    "description": "Mensagem de erro, se houver",
                    "type": "string",
                    "example": "Parâmetro inválido"
                },
                "message": {
                    "description": "Mensagem opcional para o usuário",
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "description": "Indica se a operação foi bem-sucedida",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}